refactor: move standards visibility/overrides persistence to data layer
Pure validation/partitioning/param-resolution stays in core/standards;
api goes through the new services/standards_prefs facade.
