+++ src/quodeq/core/standards/visibility.py
@@ -13,16 +13,12 @@ An absent file means the six ISO defaults, NOT "everything". Quodeq ships more
 built-in standards than the default selection shows, so treating absence as
 "no filtering" would leave a fresh install disagreeing with its own dashboard.
 
-A malformed file never fails a read; it degrades to the defaults with a warning.
+A malformed file never fails a read; it degrades to the defaults with a warning.\nRead/write of the file itself lives in ``data/fs/standards_prefs.py``.
 """
 from __future__ import annotations
 
-import json
-import logging
 from collections.abc import Iterable
 from pathlib import Path
-
-_logger = logging.getLogger(__name__)
 
 VISIBILITY_RELPATH = Path(".quodeq") / "standards-visibility.json"
 
@@ -35,7 +31,7 @@ DEFAULT_VISIBLE_STANDARDS: tuple[str, ...] = (
 )
 
 
-def _normalize(ids: Iterable[object]) -> tuple[str, ...]:
+def normalize_ids(ids: Iterable[object]) -> tuple[str, ...]:
     """Lowercase, de-duplicate and drop non-strings, preserving first order."""
     out: list[str] = []
     seen: set[str] = set()
@@ -47,36 +43,6 @@ def _normalize(ids: Iterable[object]) -> tuple[str, ...]:
             seen.add(key)
             out.append(key)
     return tuple(out)
-
-
-def load_visible_standard_ids(project_root: str | Path | None) -> tuple[str, ...]:
-    """Visible standard ids for a project; defaults when absent or malformed."""
-    if not project_root:
-        return DEFAULT_VISIBLE_STANDARDS
-    path = Path(project_root) / VISIBILITY_RELPATH
-    if not path.is_file():
-        return DEFAULT_VISIBLE_STANDARDS
-    try:
-        data = json.loads(path.read_text(encoding="utf-8"))
-    except (OSError, ValueError, UnicodeDecodeError) as exc:
-        _logger.warning("Ignoring malformed standards visibility %s: %s", path, exc)
-        return DEFAULT_VISIBLE_STANDARDS
-    ids = data.get("visibleStandardIds") if isinstance(data, dict) else None
-    if not isinstance(ids, list):
-        _logger.warning(
-            "Ignoring standards visibility %s: no 'visibleStandardIds' array", path)
-        return DEFAULT_VISIBLE_STANDARDS
-    # An explicitly empty list is a real selection ("hide everything") and is
-    # returned as-is; only a missing/!list value falls back to the defaults.
-    return _normalize(ids)
-
-
-def save_visible_standard_ids(project_root: str | Path, ids: list[str]) -> None:
-    """Write the selection, creating ``.quodeq/`` when needed."""
-    path = Path(project_root) / VISIBILITY_RELPATH
-    path.parent.mkdir(parents=True, exist_ok=True)
-    payload = {"version": 1, "visibleStandardIds": list(_normalize(ids))}
-    path.write_text(json.dumps(payload, indent=2) + "\n", encoding="utf-8")
 
 
 def validate_visible_ids(raw: object, known_ids: set[str]) -> tuple[list[str], list[str]]:
@@ -95,7 +61,7 @@ def validate_visible_ids(raw: object, known_ids: set[str]) -> tuple[list[str], l
             errors.append(f"{entry}: unknown standard")
     if errors:
         return [], errors
-    return list(_normalize(raw)), []
+    return list(normalize_ids(raw)), []
 
 
 def partition_visible(
